feat: redisign signup page
